Implement container layer
