+++ web_app/client/src/component/ConcurrentViewersChart.tsx
@@ -1,0 +1,29 @@
+import { LiveDetail } from "../type";
+import React from "react";
+import { LivePresenter } from "../presenter";
+import { ChartOptions } from "chart.js";
+import { Segment } from "semantic-ui-react";
+import { Line } from "react-chartjs-2";
+
+interface Props {
+  liveDetails: LiveDetail[];
+}
+
+export const ConcurrentViewersChart: React.FC<Props> = ({liveDetails}) => {
+  const data = LivePresenter.constructConcurrentViewersData(liveDetails);
+
+  const options: ChartOptions = {
+    legend: {
+      align: "end",
+    },
+    animation: {
+      duration: 0
+    }
+  }
+
+  return (
+      <Segment>
+        <Line data={data} options={options} />
+      </Segment>
+  )
+};
