+++ web_app/client/src/component/SettingsHeader.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import { Divider, Header } from "semantic-ui-react";
+
+export const SettingsHeader: React.FC = () => (
+    <div>
+      <Header as='h3'>
+        <Header.Content>
+          設定
+        </Header.Content>
+      </Header>
+      <Divider />
+    </div>
+)
