+++ web_app/client/src/component/index.ts
@@ -1,0 +1,6 @@
+export { AppHeader } from "./AppHeader";
+export { ConcurrentViewersChart } from "./ConcurrentViewersChart"
+export { DashboardHeader } from "./DashboardHeader";
+export { LikeAndDislikeChart } from "./LikeAndDislikeChart";
+export { LiveInformation } from "./LiveInformation";
+export { Dashboard } from "./Dashboard";
