+++ web_app/client/src/container/index.ts
@@ -1,0 +1,3 @@
+export { Dashboard } from "./Dashboard";
+export { AppHeader } from "./AppHeader";
+export { Settings } from "./Settings";
